Added End of Fight EXP Bar
Added EXP Bar to end of fights. Took forever and still not happy with it, but it'll do for now.

+++ Project Potential/Assets/Scripts/ModNumberController.cs
@@ -13,7 +13,7 @@ public class ModNumberController : MonoBehaviour {
     private int spiritMod;
     private int playerLevel;
     private int statPoints;
-    private int experiencePoints;
+    private float experiencePoints;
 
     private PlayerAbilities playerAbilities;
